chore: move the share button to topbar
