+++ packages/hedgehog-lab/src/components/Results/Results.tsx
@@ -6,6 +6,7 @@ import {
   IconButton,
   LinearProgress,
   Paper,
+  Skeleton,
   Tooltip,
   Typography,
 } from "@mui/material";
@@ -25,21 +26,14 @@ import SharePopup from "../Share/SharePopup";
 const Results = (): React.ReactElement => {
   const compilerLoading = useRecoilValue<boolean>(compilerLoadingState);
   const compilerResult = useRecoilValue<any>(compilerResultState);
-  const [resultFullScreen, setResultFullScreen] = useRecoilState<boolean>(
-    resultFullScreenState
-  );
 
   const { outputString, outputItem } = compilerResult;
-
-  const handleResultFullScreen = useCallback(() => {
-    setResultFullScreen(!resultFullScreen);
-  }, [resultFullScreen, setResultFullScreen]);
 
   return (
     <div style={{ height: "100%" }}>
       <Card
         sx={{
-          height: "calc(100vh - 160px)",
+          height: "calc(100vh - 100px)",
           overflowY: "auto",
           overflowX: "auto",
           borderRadius: 0,
@@ -47,25 +41,24 @@ const Results = (): React.ReactElement => {
         }}
       >
         {outputItem.length === 0 && outputString === "" ? (
-          <Box>{compilerLoading && <LinearProgress />}</Box>
+          <Box>
+            {compilerLoading && (
+              <>
+                {Array.from({ length: 10 }).map((_, index) => {
+                  return (
+                    <Box key={index} sx={{ mb: 3 }}>
+                      <Skeleton variant="rectangular" />
+                      <Skeleton variant="text" />
+                      <Skeleton variant="text" />
+                      <Skeleton variant="text" />
+                    </Box>
+                  );
+                })}
+              </>
+            )}
+          </Box>
         ) : (
           <Paper elevation={0} sx={{ minHeight: "100%", borderRadius: 0 }}>
-            <Box
-              sx={{
-                display: "flex",
-                justifyContent: "end",
-                position: "sticky",
-                top: 0,
-                right: 0,
-              }}
-            >
-              <SharePopup url="https://hhlab.dev/" />
-
-              <IconButton onClick={handleResultFullScreen}>
-                <FullscreenOutlined />
-              </IconButton>
-            </Box>
-
             {outputItem.length > 0 && (
               <div>
                 <Output outputItemList={outputItem} />
